requirements.txt
ray>=2.9.0
Pillow>=10.0.0
